requirements.txt
openai>=1.0
anthropic>=0.20
google-genai>=1.0
PyMuPDF>=1.24
python-docx>=1.0
rich>=13.0
tiktoken>=0.7
